Move resolve_to_key_addr to src/lib.rs
